work in prongress

+++ src/prototype/Fortran/proxG1.f90
@@ -8,8 +8,8 @@ subroutine proxG1(Umbar,Ufbar,Vm,Vf,mbar,fbar,m,f,g,N,Q)
     double precision, dimension(N+1,Q+1), intent(in) :: m,f; 
     double precision, dimension(N+2,Q+1), intent(in) :: mbar;
     double precision, dimension(N+1,Q+2), intent(in) :: fbar; 
-    real, intent(in) :: g;
-    
+    double precision, intent(in) :: g;
+
     !! La projection !!
     double precision, dimension(N+1,Q+1), intent(out) :: Vm,Vf; ! projection de m sur J 
     double precision, dimension(N+2,Q+1), intent(out) :: Umbar; ! projection de mbar sur C
